MAINT: Renamed extensions to plugins in MNE analyze
