+++ applications/mne_analyze/libs/anShared/Utils/types.h
@@ -91,7 +91,7 @@ namespace ANSHAREDLIB
     enum EVENT_TYPE
     {
         PING,                       // dummy event for testing and debuggin purposes
-        EXTENSION_INIT_FINISHED,    // send when all extensions finished initializing
+        PLUGIN_INIT_FINISHED,    // send when all plugins finished initializing
         STATUS_BAR_MSG,             // sending a message to the status bar (part of gui)
         CURRENTLY_SELECTED_MODEL,   // event send whenever the user changes the selection in the datamanager plugin
         NEW_ANNOTATION_ADDED,       // event send whenever the user adds a new annotation in the rawdataviewer plugin
