+++ applications/mne_analyze/libs/anShared/anShared.pro
@@ -80,7 +80,7 @@ CONFIG(debug, debug|release) {
 SOURCES += \
     Management/analyzedata.cpp \
     Management/analyzesettings.cpp \
-    Management/extensionmanager.cpp \
+    Management/pluginmanager.cpp \
     Management/event.cpp \
     Management/communicator.cpp \
     Management/eventmanager.cpp \
@@ -93,13 +93,13 @@ HEADERS += \
     Model/abstractmodel.h \
     Management/analyzedata.h \
     Management/analyzesettings.h \
-    Management/extensionmanager.h \
+    Management/pluginmanager.h \
     Management/event.h \
     Management/communicator.h \
     Management/eventmanager.h \
     Management/statusbar.h \
     Interfaces/IStandardView.h \
-    Interfaces/IExtension.h \
+    Interfaces/IPlugin.h \
     Utils/metatypes.h \
     Utils/types.h \
     Model/fiffrawviewmodel.h \
