+++ applications/mne_analyze/mne_analyze.pro
@@ -40,10 +40,10 @@ TEMPLATE = subdirs
 
 SUBDIRS += \
     libs \
-    extensions \
+    plugins \
     mne_analyze \
 
 # Specify dependencies because of packaging on MacOS
 libs.depends =
-extensions.depends = libs
-mne_analyze.depends = libs extensions
+plugins.depends = libs
+mne_analyze.depends = libs plugins
