+++ applications/mne_analyze/plugins/annotationmanager/annotationdelegate.cpp
