+++ applications/mne_analyze/plugins/annotationmanager/annotationdelegate.h
