+++ applications/mne_analyze/plugins/annotationmanager/annotationmanager.pro
@@ -1,15 +1,14 @@
 #==============================================================================================================
 #
-# @file     rawdataviewer.pro
+# @file     annotationmanager.pro
 # @author   Lorenz Esch <lesch@mgh.harvard.edu>;
-#           Lars Debor <Lars.Debor@tu-ilmenau.de>;
-#           Simon Heinke <Simon.Heinke@tu-ilmenau.de>
+#           Gabriel Motta <gbmotta@mgh.harvard.edu>
 # @since    0.1.0
-# @date     October, 2018
+# @date     Februrary, 2020
 #
 # @section  LICENSE
 #
-# Copyright (C) 2018, Lorenz Esch, Lars Debor, Simon Heinke. All rights reserved.
+# Copyright (C) 2020, Lorenz Esch, Gabriel Motta. All rights reserved.
 #
 # Redistribution and use in source and binary forms, with or without modification, are permitted provided that
 # the following conditions are met:
@@ -30,7 +29,7 @@
 # POSSIBILITY OF SUCH DAMAGE.
 #
 #
-# @brief    This project file generates the makefile for the rawdataviewer plugin.
+# @brief    This project file generates the makefile for the annotationmanager plugin.
 #
 #==============================================================================================================
 
@@ -40,21 +39,20 @@ TEMPLATE = lib
 
 CONFIG += plugin
 
-DEFINES += RAWDATAVIEWER_EXTENSION
+DEFINES += ANNOTATIONMANAGER_PLUGIN
 
-QT += gui widgets charts svg opengl
+QT += gui widgets
 
-contains(MNECPP_CONFIG, noOpenGL) {
-    DEFINES += NO_OPENGL
-    QT -= opengl
+contains(MNECPP_CONFIG, wasm) {
+    DEFINES += WASMBUILD
 }
 
-TARGET = rawdataviewer
+TARGET = annotationmanager
 CONFIG(debug, debug|release) {
     TARGET = $$join(TARGET,,,d)
 }
 
-DESTDIR = $${MNE_BINARY_DIR}/mne_analyze_extensions
+DESTDIR = $${MNE_BINARY_DIR}/mne_analyze_plugins
 
 contains(MNECPP_CONFIG, static) {
     CONFIG += staticlib
@@ -89,19 +87,25 @@ CONFIG(debug, debug|release) {
 }
 
 SOURCES += \
-    rawdataviewer.cpp \
-    fiffrawview.cpp \
-    fiffrawviewdelegate.cpp
+    annotationmanager.cpp \
+    annotationdelegate.cpp \
+#    annotationmodel.cpp \
+    annotationview.cpp \
+#    FormFiles/annotationmanagercontrol.cpp
 
 HEADERS += \
-    rawdataviewer_global.h \
-    rawdataviewer.h    \
-    fiffrawview.h \
-    fiffrawviewdelegate.h
+    annotationmanager_global.h \
+    annotationmanager.h \
+    annotationdelegate.h \
+#    annotationmodel.h \
+    annotationview.h \
+#    FormFiles/annotationmanagercontrol.h
 
 FORMS += \
+    annotationview.ui \
+#    FormFiles/annotationmanagercontrol.ui
 
-OTHER_FILES += rawdataviewer.json
+OTHER_FILES += annotationmanager.json
 
 RESOURCES += \
 
@@ -112,7 +116,7 @@ COPY_CMD = $$copyResources($${RESOURCE_FILES})
 QMAKE_POST_LINK += $${COPY_CMD}
 
 # Put generated form headers into the origin --> cause other src is pointing at them
-UI_DIR = $$PWD
+UI_DIR = $${PWD}
 
 INCLUDEPATH += $${EIGEN_INCLUDE_DIR}
 INCLUDEPATH += $${MNE_INCLUDE_DIR}
@@ -120,7 +124,7 @@ INCLUDEPATH += $${MNE_ANALYZE_INCLUDE_DIR}
 
 # Install headers to include directory
 header_files.files = $${HEADERS}
-header_files.path = $${MNE_INSTALL_INCLUDE_DIR}/mne_analyze_extensions
+header_files.path = $${MNE_INSTALL_INCLUDE_DIR}/mne_analyze_plugins
 
 unix: QMAKE_CXXFLAGS += -isystem $$EIGEN_INCLUDE_DIR
 
@@ -132,8 +136,8 @@ unix:!macx {
     QMAKE_RPATHDIR += $ORIGIN/../../lib
 }
 
-# Activate FFTW backend in Eigen
-contains(MNECPP_CONFIG, useFFTW) {
+# Activate FFTW backend in Eigen for non-static builds only
+contains(MNECPP_CONFIG, useFFTW):!contains(MNECPP_CONFIG, static) {
     DEFINES += EIGEN_FFTW_DEFAULT
     INCLUDEPATH += $$shell_path($${FFTW_DIR_INCLUDE})
     LIBS += -L$$shell_path($${FFTW_DIR_LIBS})
