+++ applications/mne_analyze/plugins/annotationmanager/annotationview.cpp
