+++ applications/mne_analyze/plugins/annotationmanager/annotationview.h
