+++ applications/mne_analyze/plugins/dataloader/dataloader_global.h
@@ -46,7 +46,7 @@
 // PREPROCESSOR DEFINES
 //=============================================================================================================
 
-#if defined(DATALOADER_EXTENSION)
+#if defined(DATALOADER_PLUGIN)
 #  define DATALOADERSHARED_EXPORT Q_DECL_EXPORT   /**< Q_DECL_EXPORT must be added to the declarations of symbols used when compiling a shared library. */
 #else
 #  define DATALOADERSHARED_EXPORT Q_DECL_IMPORT   /**< Q_DECL_IMPORT must be added to the declarations of symbols used when compiling a client that uses the shared library. */
