+++ applications/mne_analyze/plugins/datamanager/FormFiles/datamanagerview.cpp
