+++ applications/mne_analyze/plugins/datamanager/datamanager.h
@@ -1,13 +1,15 @@
 //=============================================================================================================
 /**
- * @file     dataloader.h
- * @author   Lorenz Esch <lesch@mgh.harvard.edu>
+ * @file     datamanager.h
+ * @author   Lorenz Esch <lesch@mgh.harvard.edu>;
+ *           Lars Debor <Lars.Debor@tu-ilmenau.de>;
+ *           Simon Heinke <Simon.Heinke@tu-ilmenau.de>
  * @since    0.1.0
- * @date     November, 2019
+ * @date     August, 2018
  *
  * @section  LICENSE
  *
- * Copyright (C) 2019, Lorenz Esch. All rights reserved.
+ * Copyright (C) 2018, Lorenz Esch, Lars Debor, Simon Heinke. All rights reserved.
  *
  * Redistribution and use in source and binary forms, with or without modification, are permitted provided that
  * the following conditions are met:
@@ -28,20 +30,19 @@
  * POSSIBILITY OF SUCH DAMAGE.
  *
  *
- * @brief    Contains the declaration of the DataLoader class.
+ * @brief    Contains the declaration of the DataManager class.
  *
  */
 
-#ifndef DATALOADER_H
-#define DATALOADER_H
+#ifndef LOADEDDATA_H
+#define LOADEDDATA_H
 
 //=============================================================================================================
 // INCLUDES
 //=============================================================================================================
 
-#include "dataloader_global.h"
-
-#include <anShared/Interfaces/IExtension.h>
+#include "datamanager_global.h"
+#include <anShared/Interfaces/IPlugin.h>
 
 //=============================================================================================================
 // QT INCLUDES
@@ -50,86 +51,81 @@
 #include <QtWidgets>
 #include <QtCore/QtPlugin>
 #include <QDebug>
+#include <QPointer>
 
 //=============================================================================================================
 // FORWARD DECLARATIONS
 //=============================================================================================================
+
+class DataManagerView;
 
 namespace ANSHAREDLIB {
     class Communicator;
 }
 
 //=============================================================================================================
-// DEFINE NAMESPACE DATALOADEREXTENSION
+// DEFINE NAMESPACE SURFERPLUGIN
 //=============================================================================================================
 
-namespace DATALOADEREXTENSION
+namespace DATAMANAGERPLUGIN
 {
-
-//=============================================================================================================
-// DATALOADEREXTENSION FORWARD DECLARATIONS
-//=============================================================================================================
-
-class DataLoaderControl;
 
 //=============================================================================================================
 /**
- * DataLoader Extension
+ * DataManager Plugin
  *
- * @brief The DataLoader class provides input and output capabilities for the fiff file format.
+ * @brief The DataManager class provides a view with all currently loaded models.
  */
-class DATALOADERSHARED_EXPORT DataLoader : public ANSHAREDLIB::IExtension
+class DATAVIEWERSHARED_EXPORT DataManager : public ANSHAREDLIB::IPlugin
 {
     Q_OBJECT
-    Q_PLUGIN_METADATA(IID "ansharedlib/1.0" FILE "dataloader.json") //New Qt5 Plugin system replaces Q_EXPORT_PLUGIN2 macro
+    Q_PLUGIN_METADATA(IID "ansharedlib/1.0" FILE "datamanager.json") //New Qt5 Plugin system replaces Q_EXPORT_PLUGIN2 macro
     // Use the Q_INTERFACES() macro to tell Qt's meta-object system about the interfaces
-    Q_INTERFACES(ANSHAREDLIB::IExtension)
+    Q_INTERFACES(ANSHAREDLIB::IPlugin)
 
 public:
     //=========================================================================================================
     /**
-     * Constructs a DataLoader.
+     * Constructs a DataManager.
      */
-    DataLoader();
+    DataManager();
 
     //=========================================================================================================
     /**
-     * Destroys the DataLoader.
+     * Destroys the DataManager.
      */
-    ~DataLoader() override;
+    virtual ~DataManager() override;
 
-    // IExtension functions
-    virtual QSharedPointer<IExtension> clone() const override;
+    // IPlugin functions
+    virtual QSharedPointer<IPlugin> clone() const override;
     virtual void init() override;
     virtual void unload() override;
     virtual QString getName() const override;
-
     virtual QMenu* getMenu() override;
     virtual QDockWidget* getControl() override;
     virtual QWidget* getView() override;
-
     virtual void handleEvent(QSharedPointer<ANSHAREDLIB::Event> e) override;
     virtual QVector<ANSHAREDLIB::EVENT_TYPE> getEventSubscriptions() const override;
 
 private:
     //=========================================================================================================
     /**
-     * This functions is called when the load from file button is pressed.
+     * Updates the content of the list widget.
      */
-    void onLoadFiffFilePressed();
+    void updateListWidget();
 
-    //=========================================================================================================
-    /**
-     * This functions is called when the save to file button is pressed.
-     */
-    void onSaveFiffFilePressed();
+    void onCurrentlySelectedModelChanged(const QString& sCurrentModelPath);
 
+    // Control
+    QPointer<QDockWidget>               m_pControlDock;         /**< Control Widget */
+    QPointer<DataManagerView>           m_pDataManagerView;
     QPointer<ANSHAREDLIB::Communicator> m_pCommu;
-
-    QString                             m_sCurrentlySelectedModel;
-
 };
+
+//=============================================================================================================
+// INLINE DEFINITIONS
+//=============================================================================================================
 
 } // NAMESPACE
 
-#endif // DATALOADER_H
+#endif // LOADEDDATA_H
