+++ applications/mne_analyze/plugins/rawdataviewer/fiffrawview.cpp
@@ -67,7 +67,7 @@
 // USED NAMESPACES
 //=============================================================================================================
 
-using namespace RAWDATAVIEWEREXTENSION;
+using namespace RAWDATAVIEWERPLUGIN;
 using namespace ANSHAREDLIB;
 
 //=============================================================================================================
