+++ applications/mne_analyze/plugins/rawdataviewer/fiffrawviewdelegate.cpp
@@ -60,7 +60,7 @@
 // USED NAMESPACES
 //=============================================================================================================
 
-using namespace RAWDATAVIEWEREXTENSION;
+using namespace RAWDATAVIEWERPLUGIN;
 using namespace ANSHAREDLIB;
 
 //=============================================================================================================
